+++ src/main/java/com/rootcode/practicalevaluation/services/user/UserService.java
@@ -5,6 +5,6 @@ import com.rootcode.practicalevaluation.dto.user.UserDTO;
 import com.rootcode.practicalevaluation.services.CrudService;
 import org.springframework.security.core.userdetails.UserDetailsService;
 
-public interface UserService extends UserDetailsService, CrudService<UserDTO, Long> {
+public interface UserService extends UserDetailsService, CrudService<UserDTO, Integer> {
     StandardResponse getBorrowingHistory(String userEmail);
 }
