+++ src/main/java/com/rootcode/practicalevaluation/services/user/impl/UserServiceImpl.java
@@ -74,7 +74,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public StandardResponse findById(Long id) {
+    public StandardResponse findById(Integer id) {
         var user = userRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("User not found with ID: " + id));
 
@@ -83,7 +83,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public StandardResponse deleteById(Long id) {
+    public StandardResponse deleteById(Integer id) {
         if (!userRepository.existsById(id)) {
             throw new ResourceNotFoundException("User not found with ID: " + id);
         }
